UPBGE: Disable objects LoD while we render cubemaps by default
waiting we do performances tests to see if LOD can be really efficient
while we render cubemaps, I disable it by default. An option is created in
scene tab to enable it.

We need to check if looping through all objects in the scene 6 times when
we render an unique cubemap can't have a bad performance impact.