BGE: KX_GameObject.refine(cutobj,accuracy) to cut soft body.
The function has only an effect on a soft body game object. The cutobj
must be a mesh object with a single face (tri or quad) that will be used
as a cut device. All the links and faces that are touching the cut object
will be cut. It can be partial cut or a full cut (the softbody cut in 2
separate parts).
accuracy is a floating point value representing the precision of the cut:
edges touching the cut plane within this distance won't be refined,
causing slightly zig-zaging cut line. Avoid using too small accuracy,
it yields very small faces during the refine process that make the
softbody simulation less stable.

Return value is True if an effective cut was done, False otherwise.

This is a first release with the following limitations:
- the soft body must have a single material
- the soft body must be smooth (not flat face)