BGE softbody cutting: new rest length derived from old rest length.
Previously, the rest length were recomputed from the current shape
of the softbody, which was incorrect as it caused the softbody to
relax if it was under tension.
The modified links will now get their rest length computed from
the other links original rest length by application of the cosine law.