BGE: Fix crash with stereo
When stereo is activated, dbvt culling is disabled and thus all objects are
set to render, including lamps. Lamps don't have meshes but the UpdateBucket()
function was not checking that.

Note: dbvt culling is disabled with stereo because the visibility of an object
could be different between eyes when it sits on the border of the frustrum.