BGE: Fix stereo bug causing crash.
Stereo mode requires a temp camera. It was incorrectly released
before being stored in CameraRenderData => pointer to freed memory!

Notes: previous commit fixes the no-culling mode, which was indeed buggy,
but the fact that culling was disabled in stereo was not voluntary; it
was the consequence of memory corruption caused by this bug.