BGE: control deformer normal recalculation via property __NORMALS__
A game object with a modifier (armature or shape keys) will skip normal
recalculation if the property __NORMALS__ is defined with value 0 on the
object. The property can be boolean, integer or float.
If the property is not defined, the normal are recalculated by default.

Note: the property is only checked at game start; changing the property value
during the game will have no effect. Use the KX_GameObject.setRecalcNormal()
python function for that.