UPBGE: Fix compound shape hierarchy.
Previously parent object of a compound assembly must have been
the root object of the parent hierarchy, this caused that if the user
want to compound assembly parent to an object without physics it was
impossible to use.

To fix so we don't refers to the root parent but the upper parent
object with compound shape option.